backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Email validation
email-validator==2.1.0

# File handling and utilities
python-magic==0.4.27
aiofiles==23.2.1

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# ML/AI dependencies (for future sprints)
sentence-transformers==2.2.2
faiss-cpu==1.7.4
groq==0.4.1

# PDF processing (for future sprints)
PyPDF2==3.0.1
python-docx==1.1.0

# Data processing
pandas==2.1.4
numpy==1.24.4
